`clippy_dev`: Move some things from `parse` to `utils`.
